requirements.txt
biopython==1.77
certifi==2020.4.5.1
joblib==0.15.1
mmh3==2.5.1
pandas==1.0.3
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
threadpoolctl==2.1.0
tqdm==4.46.1
